+++ crates/iceberg-catalog/src/api/management/v1/user.rs
@@ -112,7 +112,7 @@ pub struct ListUsersQuery {
     /// Signals an upper bound of the number of results that a client will receive.
     /// Default: 100
     #[serde(default = "default_page_size")]
-    pub page_size: i32,
+    pub page_size: i64,
 }
 
 impl ListUsersQuery {
